src/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
markdown==3.5.1
python-dotenv==1.0.0
sqlalchemy==2.0.41